+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## `master`
 
 - Cache only 2xx responses, closes #104
+- Set caching headers on statics, closes #103
 
 ## `0.70.3`